unmove some stuff for a clearer diff
