+++ src/go/cmd/api_view_test.go
@@ -286,12 +286,12 @@ func TestVars(t *testing.T) {
 	require.True(t, hasHTTPClient)
 }
 
-func TestGetReviewNameFromModPath(t *testing.T) {
-	require.EqualValues(t, "foo", getReviewNameFromModPath("foo"))
-	require.EqualValues(t, "foo", getReviewNameFromModPath("foo/v2"))
-	require.EqualValues(t, "sdk/foo", getReviewNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo"))
-	require.EqualValues(t, "sdk/foo/bar", getReviewNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo/bar"))
-	require.EqualValues(t, "sdk/foo/bar", getReviewNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo/bar/v5"))
+func Test_getPackageNameFromModPath(t *testing.T) {
+	require.EqualValues(t, "foo", getPackageNameFromModPath("foo"))
+	require.EqualValues(t, "foo", getPackageNameFromModPath("foo/v2"))
+	require.EqualValues(t, "sdk/foo", getPackageNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo"))
+	require.EqualValues(t, "sdk/foo/bar", getPackageNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo/bar"))
+	require.EqualValues(t, "sdk/foo/bar", getPackageNameFromModPath("github.com/Azure/azure-sdk-for-go/sdk/foo/bar/v5"))
 }
 
 func TestDeterministicOutput(t *testing.T) {
